APA102 Example - set kApaMatrixWidth/kApaMatrixHeight to reasonable types for a long LED strip

+++ examples/FastLed_Panel_Plus_Apa102/FastLed_Panel_Plus_Apa102.ino
@@ -49,8 +49,8 @@ SMARTMATRIX_ALLOCATE_SCROLLING_LAYER(scrollingLayer, kMatrixWidth, kMatrixHeight
 #endif
 
 #if (ENABLE_APA102_REFRESH == 1)
-const uint8_t kApaMatrixWidth = 8;          // adjust this to your APA matrix/strip
-const uint8_t kApaMatrixHeight = 8;         // set kApaMatrixHeight to 1 for a strip
+const uint16_t kApaMatrixWidth = 8;          // adjust this to your APA matrix/strip
+const uint16_t kApaMatrixHeight = 8;         // set kApaMatrixHeight to 1 for a strip
 const uint8_t kApaRefreshDepth = 36;        // not used for APA matrices as of now
 const uint8_t kApaDmaBufferRows = 1;        // not used for APA matrices as of now
 const uint8_t kApaPanelType = 0;            // not used for APA matrices as of now
